datetime(changelog): utils move to lib as dateUtils. Imports adjusted
