+++ src/modules/Datetime/Month.js
@@ -2,7 +2,7 @@ import _ from 'lodash'
 import cx from 'classnames'
 import React, { Component, PropTypes } from 'react'
 import DayCell from './DayCell'
-import * as utils from './utils'
+import * as utils from '../../lib/dateUtils'
 
 import {
   childrenUtils,
